Change the plugin system
Before, the plugin system would work when the file name of the plugin was spoken. This backfired when I made everything required to be called main. So, this time, you can tell it what word/words it should activate on, no matter the name.

The only downside to this is that it will read multiple files in the same folder, so plugins can only have one folder for all of their operations. I haven't tried using classes yet, but they might work.