Fix a line in the setup.bat script
We are very close to releasing 2.0, and so I wanted to double check everything, starting with scripts. However, I noticed so many things wrong with it. This is just fixing the setup.bat file, but I also want to cut things from other scripts. For example, I am considering putting plugins in here by default, rather than having to install them. (VOSK and all the other programs would still be in the script because they are too large to include here.)